+++ src/components/mod.rs
@@ -1,5 +1,7 @@
+mod b_dropdown;
 mod b_modal;
 mod b_navbar;
 
+pub use b_dropdown::*;
 pub use b_modal::*;
 pub use b_navbar::*;